Made href updating logic more DRY.
