+++ packages/stateful/hooks/index.ts
@@ -6,5 +6,6 @@ export * from './useInbox'
 export * from './useLoadingDaoCardInfos'
 export * from './usePinnedDaos'
 export * from './useProfile'
+export * from './useUpdateNavigatingHref'
 export * from './useVotingModule'
 export * from './useWalletProfile'
